Updated isAudio and isVideo enabled
